fix: 초기화 되는 문제 해결 && 삭제하지 않고 누적하며 redis에 새로 업데이트된 값 저장

+++ src/main/java/com/economy/community/repository/PostRepositoryImpl.java
@@ -124,7 +124,7 @@ public class PostRepositoryImpl implements PostRepositoryCustom {
     }
 
     @Override
-    public void updatePostViewCount(Long id, int increment) {
+    public void updatePostViewCount(Long id, Long increment) {
         QPost post = QPost.post;
 
         new JPAUpdateClause(entityManager, post)
